gpu_collector: GPU ECC (Un-)Correctable
Support fetching GPU memory ECC (un-)correctable data.

This required a bit of refactoring of prior gpu_collector code primarily
around labels, as well as adding an interface to pull out specific
functionality in support of testing.

The ProcessorMemory type is concrete, and could not be easily mocked
otherwise (the alternative is testing against a live/faked Redfish server).