chassis_collector: Leak Detection health metric
Adds leak detector health.
In our very limited slice of hardware where this is available, we
discovered a Redfish response which is at odds with what gofish expects
of a thermal subsystem leak detection.

So works around that with a sort of 'fallback' behavior which teases out
leak detectors in the absence of a fully-compliant Redfish LeakDetection
collection.