config: begin adoption of modules
Similar to the blackbox_exporter, we want to support a
user-defined list of modules that the redfish_exporter may support.

Begins along on that path, by:

- Refactoring "reading a config" so it is less coupled to a
file. Take instead an io.Reader and use that in YAML unmarshal, so we
can write bits of config directly in tests. (or *shudders*
config-via-remote-thing-that-can-be-read, some day)
- Implement 'default' values, which we know we'll need soon, following
how blackbox_exporter does it via UnmarshalYAML-ing. As a treat, begin
with a default value around collection (context) deadline which is not
wired down into anything yet, but will be in following commits
- Writes some quick tests around assumptions on user-defined config