collector: Return errors from NewFooCollector
Follows up previous commit, refactoring NewFooCollector functions such
that they may return errors. Existing collectors do not, so it's a lot
of nil return there. This is fine.

Also refactors the use of these calls in collector.go as well as
main.go, if a collector fails to create they skip it.