main: emit 'meta' metrics
Under the collector's default registry and HTTP path, begin emitting two
timeseries:

- redfish_exporter_config_collector_last_status, which tracks the last
configuration validity of a module by name. In some rare case, if a
configuration is invalid this should be '0'. Otherwise, would expect it
to be '1'.
- redfish_exporter_unknown_modules_requested_total, a counter tracking
how many times incoming requests have asked for a module by name which
is unknown to the configuration map/file/what-have-you.