collector/json: refactor data caching
During PR feedback, a preference was expressed to have Collect be the
only Redfish API caller.

This is unfortunatley not possible as Describe is (well, should be) the
first-invoked function of a Prometheus registry (Describe should be
called before Collect ever is).

So there's a dynamic component to this; we cannot Describe at
compile-time. We also want Collect to operate on the same data Describe
knew about previously.

As a result, still fetch and cache data but move it out of the
NewJSONCollector call and into Describe or Collect, and use a sync.Once for
consistency.