collector: inject the module name
Only use right now is the JSON collector which returns a 'meta' series
for a module.

If passing multiple query params like (/redfish?module=foo&module=bar),
the meta series is emitted for both which is a Prometheus spec violation
and fails a scrape.

Injecting and using the module name ensures adequate uniqueness.