(feat): Make Collectors context-aware
One thing we've learned is that there's a high chance for piling up
goroutines.
This is made worse when a client might disconnect (e.g. Prometheus hits
a scrape timeout), makes a new request, meanwhile the collector is still
trying to chew through work from the now-disconnected client.

This commit helps, by making collectors context-aware, and passing in
the HTTP request context.

UNFORTUNATELY... This is a deviation in behavior. The prometheus
Collector interface does not accept a context, and while there is
interest (see https://github.com/prometheus/client_golang/issues/1538),
it is not available today.

I am not sure a better way to do this. Right now, all of the 'meat' of a
Collect() function is instead made private, and either
CollectWithContext() or Collect() may be used; the latter using a
context.TODO().

Refactors the aggregate RedfishCollector to use this new interface.