(refactor/gpu) begin spike on gpu_collector
Context woes aside, one of the biggest issues identified in #45 is a
goroutine explosion, primarily (though not limited) to the GPU collector.

Currently, the collector spawns many goroutines and hammers the gofish
client to gather metrics, leading to the above. We can reduce both by
using the direct gofish client coupled with expansion to instead get
'all' NVLink data per-GPU.

This commit is quite raw; it fails linting and testing but leads to a
functional reduction of between 2-3x in scrape durations.

The above also results in a huge reduction to spawned goroutines, from
~50 -> ~18.

Lastly, some choices were made in-situ with alignment from the team
about:
- Consolidating "gpu health/state", removing one series
- Removing impossible GPU telemetry (threads/cores)
- Removing the 'hostname' label, which is to-be-done

Further commits will clean this up iteratively and address the hostname deal.