(refactor/gpu) refactor filtering
It is possible to reduce cyclomatic complexity by gathering all GPUs up
front. We already did that, but retained some amount of nested looping
when it was not strictly required.

Telemetry only needs to know the system name or ID, so create a type
which embeds the redfish.Processor and also includes the system name and
ID that we need.

Adds a missing test for the filter, and also removes a test that was
blocking tests and covered now-gone code, anyways.