(refactor/gpu) rm most tests, expand util
- Remove most of the gpu_collector test logic, and Fail() temporarily. A
more complete solution is to identify what behavior still exists and
should be tested at the end of this refactor.
- Begin making the testutil mock/test/what-have-you server entirely
driven by isolated filesystem contents. This had indirectly come up in
an earlier PR, and we know we want a path forward where the 'dump' of a
RF system is version controlled, and doing so by a filesystem structure
is the most natural compared to anything else we could be doing. Adds a
simple httptest server with some basic middleware support, wires that up
to a given os.Root, and does the thing.