(refactor/gpu) use logger, add hash file resp
- Replace a slog reference in newer test with the test logger
- Adds testutil server capability for looking up files corresponding to
an md5 sum of the incoming request URL. Handling of incoming request URL
_including query params_ is necessary as certain paths of the GPU
collector will now use expansion, and there's a lot of
possibly-unsafe-for-shell characters in those URLS. So just md5 sum the
incoming URL string, and if a matching 'hashredresponse/<foo>.json'
exists, serve that.
- Adds hashed response files representing a GB300 NVLink metrics
response, when expanded as the collector requests.