(refactor/gpu) guard against context errors
Instead of plumbing down context super far when it's not _really_
necessary, instead have a convenience wrapper on the GPUCollector,
called 'whileContextAllows'. As long as the context is non-errored,
iterate some number of functions and invoke them.

Refactor the GPUCollector.collect() to use this wrapper for its
individual telemetry concerns.